==> analysis/03_analysis.Rmd <==
---
title: "Analysis and results"
author: "Andrew Heiss, Marc Dotson, and Suparna Chaudhry"
date: "Last run: `r format(Sys.time(), '%F')`"
output: 
  html_document:
    code_folding: hide
    pandoc_args:
      - "--default-image-extension=png"
editor_options: 
  chunk_output_type: console
---

# Load packages

```{r load-packages, warning=FALSE, message=FALSE}
library(tidyverse)
library(fastDummies)
library(rstan)
library(bayesplot)
library(tidybayes)
library(loo)
library(ggridges)
library(ggraph)
library(ggdag)
library(here)

# General settings
source(here("analysis", "options.R"))

# Make all the randomness reproducible
set.seed(1234)

# Set Stan options
options(mc.cores = parallel::detectCores())
rstan_options(auto_write = TRUE)
```


# Model calibration

After selecting the model we'd like to run, model calibration proceeds as follows. Note that model calibration is run in a separate R script given how long it takes to run each model.

```{r model-calibration, eval=FALSE}
intercept <- 0                            # Intercept-only
interactions <- 0                         # Include interactions with the intercept-only model?
public_affairs <- 1                       # Public affairs
political_ideology <- 0                   # Political ideology
social_views <- 0                         # Social views
charity_voluntarism <- 0                  # Charity and voluntarism
demographics <- 0                         # Demographics
public_political <- 0                     # Public affairs + Political ideology
public_political_social <- 0              # + Social views
public_political_social_charity <- 0      # + Charity and voluntarism
public_political_social_charity_demo <- 0 # + Demographics

# Load data and design
final_data <- read_rds(here::here("data", "derived_data", "final_data.rds"))
X <- read_rds(here::here("data", "derived_data", "final_design.rds"))
if (interactions == 1) X <- read_rds(here::here("data", "derived_data", "final_interactions.rds"))

# # ##############################
# X_new <- array(NA, dim = c(dim(X)[1:3], 21)) # 13 + interactions
# for (r in 1:dim(X)[1]) {
#   for (s in 1:dim(X)[2]) {
#     X_new[r,s,,] <- X[r,s,,c(1:13, 56:63)] # 13 + interactions
#   }
# }
# X <- X_new
# # ##############################

# Restructure choice data Y
Y <- final_data %>%
  select(contains("Q4")) %>%
  as.matrix()

# Restructure covariate matrix Z
if (intercept == 1) {
  Z <- matrix(data = 1, nrow = dim(X)[1], ncol = 1)
}
if (public_affairs == 1) {
  Z <- final_data %>%
    select(
      Q2.1, Q2.2, Q2.3_1:Q2.3_7, Q2.4, Q5.7 # Public affairs
    )
}
if (political_ideology == 1) {
  Z <- final_data %>%
    select(
      Q5.2 # Political ideology
    )
}
if (social_views == 1) {
  Z <- final_data %>%
    select(
      Q5.6, Q5.11, Q5.8, Q5.9, Q5.10 # Social views
    )
}
if (charity_voluntarism == 1) {
  Z <- final_data %>%
    select(
      Q2.7, Q2.8, Q2.5, Q2.6, Q2.9, Q2.10, Q5.4, Q5.5, Q5.3_1:Q5.3_10 # Charity and voluntarism
    )
}
if (demographics == 1) {
  Z <- final_data %>%
    select(
      Q5.12, Q5.13, Q5.14, Q5.15, Q5.16_1:Q5.16_6, Q5.17 # Demographics
    )
}
if (public_political == 1) {
  Z <- final_data %>%
    select(
      Q2.1, Q2.2, Q2.3_1:Q2.3_7, Q2.4, Q5.7, # Public affairs
      Q5.2                                   # Political ideology
    )
}
if (public_political_social == 1) {
  Z <- final_data %>%
    select(
      Q2.1, Q2.2, Q2.3_1:Q2.3_7, Q2.4, Q5.7, # Public affairs
      Q5.2,                                  # Political ideology
      Q5.6, Q5.11, Q5.8, Q5.9, Q5.10         # Social views
    )
}
if (public_political_social_charity == 1) {
  Z <- final_data %>%
    select(
      Q2.1, Q2.2, Q2.3_1:Q2.3_7, Q2.4, Q5.7,                           # Public affairs
      Q5.2,                                                            # Political ideology
      Q5.6, Q5.11, Q5.8, Q5.9, Q5.10,                                  # Social views
      Q2.7, Q2.8, Q2.5, Q2.6, Q2.9, Q2.10, Q5.4, Q5.5, Q5.3_1:Q5.3_10  # Charity and voluntarism
    )
}
if (public_political_social_charity_demo == 1) {
  Z <- final_data %>%
    select(
      Q2.1, Q2.2, Q2.3_1:Q2.3_7, Q2.4, Q5.7,                           # Public affairs
      Q5.2,                                                            # Political ideology
      Q5.6, Q5.11, Q5.8, Q5.9, Q5.10,                                  # Social views
      Q2.7, Q2.8, Q2.5, Q2.6, Q2.9, Q2.10, Q5.4, Q5.5, Q5.3_1:Q5.3_10, # Charity and voluntarism
      Q5.12, Q5.13, Q5.14, Q5.15, Q5.16_1:Q5.16_6, Q5.17               # Demographics
    )
}

# Replace factor labels with levels
Z <- Z %>%
  mutate_if(is.factor, as.integer) %>%
  mutate_if(is.integer, as.factor)

# Pad the variables names with _
colnames(Z) <- str_c(colnames(Z), "_")

# Use the output of lm to get a dummy-coded version of the covariate matrix Z
Z$y <- matrix(rnorm(nrow(Z)), ncol = 1)
out <- lm(y ~ ., Z, x = TRUE)

# Save out Z as a matrix, including an intercept
Z <- tibble(intercept = rep(1, dim(X)[1])) %>%
  bind_cols(as_tibble(out$x[,-1])) %>%
  as.matrix()

# ####################################################################################
# # if (intercept == 0) {
#   # OLD...
#   # Replace factor labels with levels
#   Z_old <- Z %>%
#   # Z <- Z %>%
#     mutate_if(is.factor, as.integer) %>%
#     mutate_if(is.integer, as.factor)
#   
#   # Pad the variables names with _
#   # colnames(Z) <- str_c(colnames(Z), "_")
#   colnames(Z_old) <- str_c(colnames(Z_old), "_")
#   
#   # Use the output of lm to get a dummy-coded version of the covariate matrix Z
#   # Z$y <- matrix(rnorm(nrow(Z)), ncol = 1)
#   # out <- lm(y ~ ., Z, x = TRUE)
#   Z_old$y <- matrix(rnorm(nrow(Z_old)), ncol = 1)
#   out <- lm(y ~ ., Z_old, x = TRUE)
#   
#   # Save out Z as a matrix, including an intercept
#   Z_old <- tibble(intercept = rep(1, dim(X)[1])) %>%
#   # Z <- tibble(intercept = rep(1, dim(X)[1])) %>%
#     bind_cols(as_tibble(out$x[,-1])) %>%
#     as.matrix()
#   
#   # NEW...
#   # Replace factor labels with levels and NAs with 0s
#   Z_new <- Z %>%
#     mutate_if(is.factor, as.integer) %>%
#     mutate_if(is.integer, as.factor) %>% 
#     replace(is.na(.), 0) %>% 
#     # Use dummy_cols() to get a dummy-coded version of factors
#     dummy_cols(
#       remove_first_dummy = TRUE,
#       remove_selected_columns = TRUE
#     ) %>% 
#     # Include an intercept, sort columns numerically, save as matrix
#     bind_cols(tibble(intercept = rep(1, dim(X)[1]))) %>% 
#     relocate(intercept, sort(colnames(.))) %>% 
#     as.matrix()
# # }
# 
# 
# 
# 
# ####################################################################################

# Specify the data for calibration in a list
data <- list(
  R = dim(X)[1],      # Number of respondents
  S = dim(X)[2],      # Number of choice tasks
  A = dim(X)[3],      # Number of choice alternatives
  I = dim(X)[4],      # Number of (estimable) attribute levels
  J = ncol(Z),        # Number of respondent-level covariates

  Gamma_mean = 0,      # Mean of population-level means
  Gamma_scale = 1,     # Scale of population-level means
  Omega_shape = 5,     # Shape of population-level scale
  tau_mean = 0,        # Mean of population-level scale
  tau_scale = 10,      # Scale of population-level scale

  Y = Y,               # Matrix of observed choices
  X = X,               # Array of experimental designs per choice task
  Z = Z                # Matrix of respondent-level covariates
)

# Run the model and save data and model output
fit <- stan(
  file = here::here("src", "stan_files", "hmnl_noncentered.stan"),
  data = data,
  iter = 10000,
  thin = 5,
  seed = 1234
)
run <- list(data = data, fit = fit)
if (intercept == 1 & interactions == 0) write_rds(run, here::here("analysis", "output", "model_runs", "intercept.rds"))
if (intercept == 1 & interactions == 1) write_rds(run, here::here("analysis", "output", "model_runs", "interactions_5e.rds"))
if (public_affairs == 1) write_rds(run, here::here("analysis", "output", "model_runs", "public_affairs.rds"))
if (political_ideology == 1) write_rds(run, here::here("analysis", "output", "model_runs", "political_ideology.rds"))
if (social_views == 1) write_rds(run, here::here("analysis", "output", "model_runs", "social_views.rds"))
if (charity_voluntarism == 1) write_rds(run, here::here("analysis", "output", "model_runs", "charity_voluntarism.rds"))
if (demographics == 1) write_rds(run, here::here("analysis", "output", "model_runs", "demographics.rds"))
if (public_political == 1) write_rds(run, here::here("analysis", "output", "model_runs", "public_political.rds"))
if (public_political_social == 1) write_rds(run, here::here("analysis", "output", "model_runs", "public_political_social.rds"))
if (public_political_social_charity == 1) write_rds(run, here::here("analysis", "output", "model_runs", "public_political_social_charity.rds"))
if (public_political_social_charity_demo == 1) write_rds(run, here::here("analysis", "output", "model_runs", "public_political_social_charity_demo.rds"))
```

The model is calibrated using a non-centered parameterization of a hierarchical multinomial logit model. Each model passed diagnostics checks, including the absence of divergent transitions, as part of model calibration.

```{stan hmnl-noncentered, output.var="hmnl_noncentered", eval=FALSE}
// Index values, hyperprior values, observations, and covariates.
data {
  int<lower = 1> R;                  // Number of respondents.
  int<lower = 1> S;                  // Number of choice tasks.
  int<lower = 2> A;                  // Number of choice alternatives.
  int<lower = 1> I;                  // Number of observation-level covariates.
  int<lower = 1> J;                  // Number of population-level covariates.

  real Gamma_mean;                   // Mean of population-level means.
  real<lower=0> Gamma_scale;         // Scale of population-level means.
  real<lower=0> Omega_shape;         // Shape of population-level scale.
  real tau_mean;                     // Mean of population-level scale.
  real<lower=0> tau_scale;           // Scale of population-level scale.

  int<lower = 1, upper = A> Y[R, S]; // Matrix of observations.
  matrix[A, I] X[R, S];              // Array of observation-level covariates.
  matrix[R, J] Z;                    // Matrix of population-level covariates.
}

// Parameters and hyperparameters.
parameters {
  matrix[J, I] Gamma;                // Matrix of population-level hyperparameters.
  corr_matrix[I] Omega;              // Population model correlation matrix hyperparameters.
  vector<lower = 0>[I] tau;          // Population model vector of scale hyperparameters.
  matrix[R, I] Delta;                // Matrix of non-centered observation-level parameters.
}

// Deterministic transformation.
transformed parameters {
  // Matrix of centered observation-level parameters.
  matrix[R, I] Beta;

  // Non-centered parameterization.
  for (r in 1:R) {
    Beta[r,] = Z[r,] * Gamma + Delta[r,] * quad_form_diag(Omega, tau);
  }
}

// Hierarchical multinomial logit model.
model {
  // Hyperpriors.
  to_vector(Gamma) ~ normal(Gamma_mean, Gamma_scale);
  Omega ~ lkj_corr(Omega_shape);
  tau ~ normal(tau_mean, tau_scale);

  // Non-centered population model and likelihood.
  for (r in 1:R) {
    Delta[r,] ~ normal(0, 1);
    for (s in 1:S) {
      Y[r, s] ~ categorical_logit(X[r, s] * Beta[r,]');
    }
  }
}

// Generated quantities conditioned on parameter draws.
generated quantities {
  // Compute log likelihood for model fit.
  matrix[R, S] log_lik;
  for (r in 1:R) {
    for (s in 1:S) {
      log_lik[r, s] = categorical_logit_lpmf(Y[r, s] | X[r, s] * Beta[r,]');
    }
  }
}
```


# Compute model fit and extract draws

The following analysis follows specifying the model we'd like to evaluate. Note that the output of the model calibration is too large to push to GitHub, but we can push the draws once we extract them.

## Load model output

```{r load-data-output, warning=FALSE, message=FALSE, eval=FALSE}
# Load all model output
intercept_run <- read_rds(here::here("analysis", "output", "model_runs", "intercept.rds"))
public_affairs_run <- read_rds(here::here("analysis", "output", "model_runs", "public_affairs.rds"))
political_ideology_run <- read_rds(here::here("analysis", "output", "model_runs", "political_ideology.rds"))
social_views_run <- read_rds(here::here("analysis", "output", "model_runs", "social_views.rds"))
charity_voluntarism_run <- read_rds(here::here("analysis", "output", "model_runs", "charity_voluntarism.rds"))
demographics_run <- read_rds(here::here("analysis", "output", "model_runs", "demographics.rds"))
public_political_run <- read_rds(here::here("analysis", "output", "model_runs", "public_political.rds"))
public_political_social_run <- read_rds(here::here("analysis", "output", "model_runs", "public_political_social.rds"))
public_political_social_charity_run <- read_rds(here::here("analysis", "output", "model_runs", "public_political_social_charity.rds"))
public_political_social_charity_demo_run <- read_rds(here::here("analysis", "output", "model_runs", "public_political_social_charity_demo.rds"))
```

## Compute and save model fit

```{r compute-fit, eval=FALSE}
# Compute fit using PSIS-LOO
intercept_loo <- loo(intercept_run$fit, save_psis = TRUE)
public_affairs_loo <- loo(public_affairs_run$fit, save_psis = TRUE)
political_ideology_loo <- loo(political_ideology_run$fit, save_psis = TRUE)
social_views_loo <- loo(social_views_run$fit, save_psis = TRUE)
charity_voluntarism_loo <- loo(charity_voluntarism_run$fit, save_psis = TRUE)
demographics_loo <- loo(demographics_run$fit, save_psis = TRUE)
public_political_loo <- loo(public_political_run$fit, save_psis = TRUE)
public_political_social_loo <- loo(public_political_social_run$fit, save_psis = TRUE)
public_political_social_charity_loo <- loo(public_political_social_charity_run$fit, save_psis = TRUE)
public_political_social_charity_demo_loo <- loo(public_political_social_charity_demo_run$fit, save_psis = TRUE)

# Compare model fit
loo_compare(intercept_loo, public_affairs_loo, political_ideology_loo, social_views_loo, charity_voluntarism_loo, demographics_loo, public_political_loo, public_political_social_loo, public_political_social_charity_loo, public_political_social_charity_demo_loo)
```

```{r save-fit, eval=FALSE}
# Create data frame
model_fit <- tibble(model = rep(NA, 10), elpd_loo = rep(NA, 10))

# Extract PSIS-LOO fit
model_fit$model[1] <- "intercept"; model_fit$loo[1] <- intercept_loo$elpd_loo
model_fit$model[2] <- "public_affairs"; model_fit$loo[2] <- public_affairs_loo$elpd_loo
model_fit$model[3] <- "political_ideology"; model_fit$loo[3] <- political_ideology_loo$elpd_loo
model_fit$model[4] <- "social_views"; model_fit$loo[4] <- social_views_loo$elpd_loo
model_fit$model[5] <- "charity_voluntarism"; model_fit$loo[5] <- charity_voluntarism_loo$elpd_loo
model_fit$model[6] <- "demographics"; model_fit$loo[6] <- demographics_loo$elpd_loo
model_fit$model[7] <- "public_political"; model_fit$loo[7] <- public_political_loo$elpd_loo
model_fit$model[8] <- "public_political_social"; model_fit$loo[8] <- public_political_social_loo$elpd_loo
model_fit$model[9] <- "public_political_social_charity"; model_fit$loo[9] <- public_political_social_charity_loo$elpd_loo
model_fit$model[10] <- "public_political_social_charity_demo"; model_fit$loo[10] <- public_political_social_charity_demo_loo$elpd_loo

# Save model output
write_rds(model_fit, here::here("analysis", "output", "tables", "model_fit.rds"))
```

## Load model fit objects and extract and save draws

```{r extract-draws, eval=FALSE}
intercept <- 1                            # Intercept-only
interactions <- 1                         # Include interactions with the intercept-only model?
public_affairs <- 0                       # Public affairs
political_ideology <- 0                   # Political ideology
social_views <- 0                         # Social views
charity_voluntarism <- 0                  # Charity and voluntarism
demographics <- 0                         # Demographics
public_political <- 0                     # Public affairs + Political ideology
public_political_social <- 0              # + Social views
public_political_social_charity <- 0      # + Charity and voluntarism
public_political_social_charity_demo <- 0 # + Demographics

# Load and extract data and model output
if (intercept == 1 & interactions == 0) run <- read_rds(here::here("analysis", "output", "model_runs", "intercept.rds"))
if (intercept == 1 & interactions == 1) run <- read_rds(here::here("analysis", "output", "model_runs", "interactions_5d.rds"))
if (public_affairs == 1) run <- read_rds(here::here("analysis", "output", "model_runs", "public_affairs.rds"))
if (political_ideology == 1) run <- read_rds(here::here("analysis", "output", "model_runs", "political_ideology.rds"))
if (social_views == 1) run <- read_rds(here::here("analysis", "output", "model_runs", "social_views.rds"))
if (charity_voluntarism == 1) run <- read_rds(here::here("analysis", "output", "model_runs", "charity_voluntarism.rds"))
if (demographics == 1) run <- read_rds(here::here("analysis", "output", "model_runs", "demographics.rds"))
if (public_political == 1) run <- read_rds(here::here("analysis", "output", "model_runs", "public_political.rds"))
if (public_political_social == 1) run <- read_rds(here::here("analysis", "output", "model_runs", "public_political_social.rds"))
if (public_political_social_charity == 1) run <- read_rds(here::here("analysis", "output", "model_runs", "public_political_social_charity.rds"))
if (public_political_social_charity_demo == 1) run <- read_rds(here::here("analysis", "output", "model_runs", "public_political_social_charity_demo.rds"))
data <- run$data
fit <- run$fit

# Extract posterior draws
draws <- fit %>% 
  spread_draws(Gamma[i, j]) %>% 
  ungroup() %>% 
  filter(.iteration > 500) %>% 
  rename(
    i = j,
    j = i
  )

# Save draws
if (intercept == 1 & interactions == 0) write_rds(draws, here::here("analysis", "output", "posterior_draws", "intercept.rds"))
if (intercept == 1 & interactions == 1) write_rds(draws, here::here("analysis", "output", "posterior_draws", "interactions_5d.rds"))
if (public_affairs == 1) write_rds(draws, here::here("analysis", "output", "posterior_draws", "public_affairs.rds"))
if (political_ideology == 1) write_rds(draws, here::here("analysis", "output", "posterior_draws", "political_ideology.rds"))
if (social_views == 1) write_rds(draws, here::here("analysis", "output", "posterior_draws", "social_views.rds"))
if (charity_voluntarism == 1) write_rds(draws, here::here("analysis", "output", "posterior_draws", "charity_voluntarism.rds"))
if (demographics == 1) write_rds(draws, here::here("analysis", "output", "posterior_draws", "demographics.rds"))
if (public_political == 1) write_rds(draws, here::here("analysis", "output", "posterior_draws", "public_political.rds"))
if (public_political_social == 1)  write_rds(draws, here::here("analysis", "output", "posterior_draws", "public_political_social.rds"))
if (public_political_social_charity == 1) write_rds(draws, here::here("analysis", "output", "posterior_draws", "public_political_social_charity.rds"))
if (public_political_social_charity_demo == 1) write_rds(draws, here::here("analysis", "output", "posterior_draws", "public_political_social_charity_demo.rds"))
```


# Original computing environment

<button data-toggle="collapse" data-target="#sessioninfo" class="btn btn-primary btn-md btn-info">Here's what we used the last time we built this page</button>

<div id="sessioninfo" class="collapse">

```{r show-session-info, echo=TRUE, width=100}
writeLines(readLines(file.path(Sys.getenv("HOME"), ".R/Makevars")))

devtools::session_info()
```

</div>
